refactor(src/main/java/constants): create a constant package and classify the constants
